+++ bot/commands/utility/listScheduled.js
@@ -1,43 +1,58 @@
 const source = require('rfr');
 // const moment = require('moment');
-
-const logger = source('bot/utils/logger');
+const cronstrue = require('cronstrue');
+// const logger = source('bot/utils/logger');
 const { sendMessage } = source('bot/utils/util');
 const agenda = source('scheduler');
 
-function scheduleMessage(message) {
+function listScheduled(message) {
     agenda.jobs({
         'data.serverID': message.guild.id,
     })
         .then((jobs) => {
-            logger.info(jobs);
-            sendMessage(message.channel, 'there are a bunch of jobs');
+            const list = jobs.filter((j) => j.attrs.repeatInterval)
+                .reduce((acc, j) => {
+                    const intervalString = cronstrue.toString(j.attrs.repeatInterval, {
+                        verbose: true,
+                    });
+                    // eslint-disable-next-line no-underscore-dangle
+                    return `${acc}\njobID: ${j.attrs._id}\nSchedule: ${intervalString}\nChannel: <#${j.attrs.data.channelID}>\nText: \`${j.attrs.data.message}\`\n--------\n`;
+                }, 'I currently have the following recurring messages scheduled:\n=================================================\n');
+            sendMessage(message.channel, `${list}`);
         });
 }
 
 module.exports = {
     args: 0,
-    name: 'scheduled-jobs',
+    name: 'list-jobs',
     botAdmin: false,
-    description: 'Schedule a recuring message to be sent to the channel that this is called from',
-    usage: '<day of the week> at <time>\n<some message to be sent>',
+    description: 'List all the scheduled jobs',
+    usage: '',
     aliases: [],
-    execute: scheduleMessage,
-    docs: `#### Schedule Message
-- Command: \`schedule\`
-- Returns: Schedules a message to be sent periodicly in the channel that the command is called from.
+    execute: listScheduled,
+    docs: `#### List scheduled jobs
+- Command: \`list-jobs\`
+- Returns: A listing of all the scheduled messages to be sent, and their intervals
 - Example usage:
 \`\`\`
 User
-> !scheduler thursdays at 6:00pm
-Hey everyeryone, remember we have DND soon!
+> !list-jobs
 
 Botomir
-> Message has been scheduled to be sent on Thurdsay March 25th at 6:00pm.
+> I currently have the following recurring messages scheduled:
+=================================================
 
-......
-Thursdat at 6:
-Botomir
-> Hey everyeryone, remember we have DND soon!
+jobID: 60adab0d980931395d457341
+Schedule: At 10:00 PM, every day
+Channel: #testing
+Text: \`test message to be sent at 10pm\`
+--------
+
+jobID: 60adab38e004013972571888
+Schedule: At 10:00 PM, every day
+Channel: #testing
+Text: \`also test message to be sent at 10pm\`
+--------
+
 \`\`\``,
 };